local-requirements.txt
numpy>=1.14
opencv-python
Flask
http://download.pytorch.org/whl/torch-0.3.0.post4-cp36-cp36m-macosx_10_7_x86_64.whl
torchvision==0.2.0
Pillow==5.0.0
gunicorn==19.7.1
scipy==1.0.0
google-cloud-storage==1.7.0